app/requirements.txt
flask==2.0.1
flask-cors==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy-serializer==1.4.1
Flask-Migrate==3.1.0
python-dotenv==0.19.2
requests==2.25.1
gunicorn==20.1.0
tornado==6.1
blinker==1.4
peewee==3.14.4
pymysql==1.0.2
